take into account user.profile_picture can be None

+++ app/api/views/user_info.py
@@ -54,12 +54,13 @@ def update_user_info():
     if "profile_picture" in data:
         if data["profile_picture"] is None:
             if user.profile_picture_id:
+                file = user.profile_picture
+                if file:
+                    File.delete(file.id)
+                    s3.delete(file.path)
+
                 user.profile_picture_id = None
                 db.session.flush()
-
-                file = user.profile_picture
-                File.delete(file.id)
-                s3.delete(file.path)
         else:
             raw_data = base64.decodebytes(data["profile_picture"].encode())
             file_path = random_string(30)
